refractor: change cursor type when hovering
